requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
aniso8601==10.0.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
Django==5.2
django-jazzmin==3.0.1
dotenv==0.9.9
Flask==3.1.0
Flask-RESTful==0.3.10
fontawesome-free==5.15.4
frozenlist==1.5.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.0
packaging==24.2
pillow==11.1.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pyOpenSSL==25.0.0
pyparsing==3.2.3
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
razorpay==1.4.2
redis==5.2.1
requests==2.32.3
rsa==4.9
service-identity==24.2.0
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.19.0
zope.interface==7.2
flask_sqlalchemy
